+++ node_package/src/ReactOnRails.full.ts
@@ -9,7 +9,10 @@ import type {
 import Client from './ReactOnRails.client';
 
 if (typeof window !== 'undefined') {
-  console.log('Optimization opportunity: "react-on-rails" includes ~14KB of server-rendering code. Browsers may not need it. See https://forum.shakacode.com/t/how-to-use-different-versions-of-a-file-for-client-and-server-rendering/1352 (Requires creating a free account)');
+  // warn to include a collapsed stack trace
+  console.warn(
+    'Optimization opportunity: "react-on-rails" includes ~14KB of server-rendering code. Browsers may not need it. See https://forum.shakacode.com/t/how-to-use-different-versions-of-a-file-for-client-and-server-rendering/1352 (Requires creating a free account). Click this for the stack trace.',
+  );
 }
 
 /**